coverage

+++ src/main/java/com/moataz/springplaygrounds/springdata/service/InstructorService.java
@@ -4,6 +4,7 @@ import com.moataz.springplaygrounds.springdata.dto.InstructorStudentCourseDTO;
 import com.moataz.springplaygrounds.springdata.dto.UUIDDTO;
 import com.moataz.springplaygrounds.springdata.entities.Instructor;
 import com.moataz.springplaygrounds.springdata.repository.InstructorRepository;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -27,15 +28,22 @@ public class InstructorService {
     }
 
     public Instructor update(Instructor instructor) {
+        if(instructorRepository.findById(instructor.getId()).isEmpty()){
+            throw new EntityNotFoundException("Instructor with id " + instructor.getId() + " does not exist");
+        }
        return instructorRepository.save(instructor);
     }
 
     public void delete(UUIDDTO UUIDDTO) {
+        if(instructorRepository.findById(UUIDDTO.getId()).isEmpty()){
+            throw new EntityNotFoundException("Instructor with id " + UUIDDTO.getId() + " does not exist");
+        }
         instructorRepository.deleteById(UUIDDTO.getId());
+
     }
 
-    public List<InstructorStudentCourseDTO> getInstructorStudentCourse() {
-        return instructorRepository.getInstructorNameAndCourses();
+    public List<InstructorStudentCourseDTO> getInstructorCourseStudent() {
+        return instructorRepository.getInstructorCourseStudent();
     }
